frontend/discord-bot/requirements.txt
discord.py>=2.3.0
aiohttp>=3.8.0
python-dotenv>=1.0.0
schedule>=1.2.0
Pillow>=10.0.0
html2text>=2020.1.16
redis>=5.0
